Ip-scanner/requirements.txt
requests
pycountry
pandas
openpyxl
python-dotenv
xlsxwriter